+++ endpoints/import.go
@@ -36,12 +36,12 @@ func HandleImport(logger *zap.SugaredLogger, terseStore storage.TerseStore) api.
 			)
 
 			// Report the error to the client.
-			code := int64(500)
+			code := 500
 			resp := &api.TerseImportDefault{Payload: &models.Error{
-				Code:    code,
+				Code:    int64(code),
 				Message: message,
 			}}
-			resp.SetStatusCode(int(code))
+			resp.SetStatusCode(code)
 			return resp
 		}
